Render out the cabal files, rather than use package.yamls
The reason we do this is that this project uses a common package.yaml that is included in multiple packages, which is actually supported in stack 2, but NOT when it happens in a dependency. Since this project is a dependency of ours, we can't have that feature in use here.

Perhaps Stack 2 will support this in the future, in which case we can
probably deprecate our fork.